Allow configuring P2P.MaxPacketMsgPayloadSize
Some chains have decided to set a MaxPacketMsgPayloadSize higher than
the tendermint@v0.34 default of 1024 bytes. This results in packets
Tenderseed rejects.

This commit adds `max_packet_msg_payload_size` to the Config so that
clients can decide on what maximum size for the message packet payload
is acceptable for their chain.