Allow configuring P2P.MaxPacketMsgPayloadSize (#16)
* Allow configuring P2P.MaxPacketMsgPayloadSize

Some chains have decided to set a MaxPacketMsgPayloadSize higher than
the tendermint@v0.34 default of 1024 bytes. This results in packets
Tenderseed rejects.

This commit adds `max_packet_msg_payload_size` to the Config so that
clients can decide on what maximum size for the message packet payload
is acceptable for their chain.

* Inherit DefaultConfig when LoadConfigFromFile

With the latest Config change, existing seed nodes will be missing
the MaxPacketMsgPayloadSize value. It will instead default to zero
for a fresh struct. That will in turn break all packet processing.

This commit pre-populates the struct by using DefaultConfig.
As result values undefined in config.toml do not break expectations.